Move MultiBadgeImageCardView to UI package
