+++ app/src/main/java/org/jellyfin/androidtv/ui/MultiBadgeImageCardView.kt
@@ -1,4 +1,4 @@
-package org.jellyfin.androidtv.presentation
+package org.jellyfin.androidtv.ui
 
 import android.animation.ObjectAnimator
 import android.content.Context
